Working start, but changing direction.
Right now, it starts up with a "choose a course"
GUI, then once a course is choosen pops up a "which players"
and "what date" GUI (though the date isn't there yet).

But I'm going to change to go right to the 2nd GUI, with a menu
for choosing the date and the course.